backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.3.26
langchain-community==0.3.27
dashscope==1.17.0
pydantic==2.11.7
python-multipart==0.0.6
requests==2.32.4
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.0